fix(signal_gen,tone): Phase 5 input validation and click suppression
5.1 - every signal_gen fill function now rejects null/zero-frame calls
and non-finite/out-of-range (0..20000 Hz) frequency or amplitude by
outputting silence instead of propagating garbage (SIG-001/SIG-003).

Also fixes wrap_phase()'s O(1) bound (SIG-002): the previous
subtract-in-a-loop form could iterate unboundedly for a large phase
increment; now uses floor()-based wrapping and resets non-finite phase
to 0 instead of propagating NaN/Inf.

5.2 - sg_piano_state_t.elapsed is now a saturating uint64_t instead of
a uint32_t that wrapped (and silently restarted the decay envelope)
after about 27 hours at 44.1 kHz (SIG-005). Validating frequency to
<=20 kHz also fixes SIG-004: the piano phase wrap's single  is only correct when dt < 1.0, which the
validated frequency range now guarantees.

5.3 - tone.c consolidates on/voice/hz/amp_pct into one mutex-guarded
tone_config_t snapshotted once per tone_fill() call, instead of four
independently-atomic fields that could be observed torn (e.g. a new
frequency paired with the old voice or amplitude) — TONE-001. Public
API (tone_set/tone_set_amplitude/tone_set_voice/tone_off/tone_get)
unchanged so web_ui/console callers are unaffected.

5.4 - adds a 10 ms linear gain ramp for on/off and amplitude-percent
transitions instead of an instant step. The generator always runs at
unity amplitude (phase keeps advancing regardless of on/off, as
before); the ramp is the only amplitude control applied in tone.c.

Several existing tone tests asserted instant on/off within a single
small block, which the new ramp intentionally no longer does; added a
tone_test_snap_gain() UNIT_TEST-only hook so those tests isolate
on/off *state* from ramp *timing*, and added two new tests that
directly verify the ramp is gradual and that it still fully converges.

Host suite: 19/19 across --strict/--strict --asan/--strict --ubsan.
Hardware-verified: boots clean, no regression.

+++ esp_i2s_source/components/signal_gen/include/signal_gen.h
@@ -41,7 +41,7 @@ void sg_sine_fill(sg_sine_state_t *st, int16_t *out, size_t frames,
  * Retrigger the envelope on each note with sg_piano_note_on(). */
 typedef struct {
     double   phase;     /* fundamental phase, [0, 2*PI) */
-    uint32_t elapsed;   /* samples since note-on (envelope clock) */
+    uint64_t elapsed;   /* samples since note-on (envelope clock); saturating */
 } sg_piano_state_t;
 
 /* Retrigger: restart the envelope (and phase) for a fresh note. */
